% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsrt.a_sample_gene_set_rank_test.R
\name{a_sample_gene_set_rank_test}
\alias{a_sample_gene_set_rank_test}
\title{Perform a rank-based single sample gene set analysis.}
\usage{
a_sample_gene_set_rank_test(x, geneset_list, alternative = "two.sided",
  test.method = "ks")
}
\arguments{
\item{x}{a sample in a named array format.}

\item{geneset_list}{a named list of gene sets, each of list element is named
and a list ofgenes.}

\item{alternative}{\code{c('less','greater','two.sided')}.
'less' ('greater') to check if the ranks of given gene set
is higher (lower) than the rest.}

\item{test.method}{\code{c('ks.test', 'wilcox.test')}.  Defaults to 'ks'.}
}
\description{
This function performs a rank-based single sample gene set analysis for
a sample (\code{x}) and a list of gene sets (\code{geneset_list}).
}
\examples{
a_sample_gene_set_rank_test(x, a_geneset_list)
}

