% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsrt.gene_set_rank_test.R
\name{gene_set_rank_test}
\alias{gene_set_rank_test}
\title{Perform a rank-based single sample gene set analysis.}
\usage{
gene_set_rank_test(data, geneset_list, alternative = "two.sided",
  test.method = "ks")
}
\arguments{
\item{data}{Either a set of samples in \code{data.frame} format
or a sample in a named array format.}

\item{geneset_list}{a named list of gene sets, each of list element is named
and a list ofgenes.}

\item{alternative}{see 'a_sample_gene_set_rank_test' for detail.}

\item{test.method}{\code{c('ks.test', 'wilcox.test')}.  Defaults to 'ks'.}
}
\description{
This function performs a rank-based single sample gene set analysis for
a set of samples (\code{data}) and a list of gene sets (\code{geneset_list}).
}
\examples{
gene_set_rank_test(data, a_geneset_list)
}

