% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsrt.get_a_value.R
\name{get_a_value}
\alias{get_a_value}
\title{Extract a specific component from gene set analysis (\code{gsa_list}) and
create a matrix.}
\usage{
get_a_value(gsa_list, which = "z.stat")
}
\arguments{
\item{gsa_list}{a list of gene set analysis from \code{gene_set_rank_test}.}

\item{which}{a name of a list element to extract. Defaults to 'z.stat'.}
}
\value{
a matrix (M gene sets x N samples) of selected element.
}
\description{
Extract a specific component from gene set analysis (\code{gsa_list}) and
create a matrix.
}

