% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsrt.parse_a_line_in_gmt.R
\name{parse_a_line_in_gmt}
\alias{parse_a_line_in_gmt}
\title{Parse a line in 'gmt' format}
\usage{
parse_a_line_in_gmt(aLine)
}
\arguments{
\item{aLine}{a string in 'gmt' format. There are assumed three main elements.
Each element is 'tab' delitmeted.  The first item is a name, the second
item is 'url' or something like that, and the last item is a list of genes,
separated by 'space'.}
}
\value{
a list of three elements: \code{name}, \code{url}, \code{geneset}
}
\description{
This function parses a line of string in 'gmt' format,
  for example a line in a MSigDB file.
}
\examples{
parse_a_line_in_gmt('a_name\\tan_url\\tgene1 gene2 gene3')
}
\author{
Seungchan Kim (dolchan@gmail.com)
}
\keyword{gmt}
\keyword{parse}

