% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsrt.read_MSigDB.R
\name{read_MSigDB}
\alias{read_MSigDB}
\title{Read a file in MSigDB/gmt format into a list}
\usage{
read_MSigDB(gmtfile, min.size = 0, max.size = -1, simplify = TRUE)
}
\arguments{
\item{gmtfile}{a text file in 'gmt' format.
See 'parse_a_line_in_gmt' function for detail.}

\item{min.size}{The minimum gene set size to read.  Defaults to 0.}

\item{max.size}{The maximum gene set size to read.  Defaults to no limit.}

\item{simplify}{Defaults to TRUE.  See 'output' for detail.}
}
\value{
a list of gene set.
       If TRUE, the return structure is simplified to a list of genesets
       each of which is a list of genes and the name of each geneset is
       the name of gene set from MSigDB file, the first element of each line.
       If FALSE, it returns a list of three element list each element of which
       is \code{name}, \code{url}, and \code{geneset}.
}
\description{
This function reads a GMT file and creates a list of gene sets.
}
\examples{
read_MSigDB('c2.cp.biocarta.v4.0.symbols.gmt.txt')
}

