% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsrt.a_sample_gene_set_rank_test.R
\name{a_sample_gene_set_rank_test}
\alias{a_sample_gene_set_rank_test}
\title{Perform a rank-based single sample gene set analysis.}
\usage{
a_sample_gene_set_rank_test(x, geneset_list, row_names = NULL,
  alternative = "two.sided", test.method = "ks")
}
\arguments{
\item{x}{a sample in a named array format.}

\item{geneset_list}{a named list of gene sets, each of list element is named
and a list ofgenes.}

\item{row_names}{a list of (gene) names of x. Useful when names(x) is NULL.}

\item{alternative}{\code{c('less','greater','two.sided')}.
If 'less'/'greater', test if the overall gene expression of given gene set
is lower/higher than the rest.  If 'two.sided', test either lower or higher.
Defaults to 'two.sided'.}

\item{test.method}{\code{c('ks.test', 'wilcox.test')}.  Defaults to 'ks'.}
}
\value{
a list of z.stat, D (ks.test) or W (wilcox.test) stat, p.value, and adj.p.value (FDR) for each gene set.
        z.stat indicates if the gene set expression is "lower" (negative) or "higher" (positive).
        We use p-value estimated from the actual statistical test (ks.test or wilcox.test) and
        use normal distribution to compute "pseudo" z statistics.
}
\description{
This function performs a rank-based single sample gene set analysis for
a sample (\code{x}) and a list of gene sets (\code{geneset_list}).
}
\examples{
a_sample_gene_set_rank_test(x, a_geneset_list)
}

